x86/vpt: fix injection to remote vCPU
vpt timers are usually added to the per-vCPU list of the vCPU where
they get setup, but depending on the timer source type that vCPU might
be different than the one where the interrupt vector gets injected.

For example the PIT timer use a PIC or IO-APIC pin in order to select
the destination vCPU and vector, which might not match the vCPU they
are configured from.

If such a situation happens pt_intr_post won't be called, and thus the
vpt will be left in a limbo where the next interrupt won't be
scheduled. Fix this by generalizing the special handling done to
IO-APIC level interrupts to be applied always when the destination
vCPU of the injected vector is different from the vCPU where the vpt
belongs to (ie: usually the one it's been configured from).

A further improvement as noted in a comment added to the code might be
to move the vpt so it's handled by the same vCPU where the vector gets
injected.

Signed-off-by: Roger Pau Monné <roger.pau@citrix.com>
---
Changes since v1:
 - Do not assigned pt_vector to the return of hvm_isa_irq_assert when
   not using the IO-APIC.
 - Expand the comment regarding the vector not being set in the lapic
   IRR register, which can mean that the IRQ/GSI has been masked at
   the PIC, or that it's been injected into a different vCPU.
 - Move the pt_vector < 0 and IRR check outside the switch.
 - Reorder some code to reduce indentation.