x86/hvm: migrate vPTs on vCPU offlining
When a vCPU is offlined the virtual periodic timers that are not tied
to a specific vCPU (those with ISA or IO-APIC sources) currently
assigned to it must be migrated, so that they can continue to fire.

In order to do so pick an online vCPU from the domain and migrate all
global vPT timers to it. This could be done in a more fine-grained
way, but then we won't be able to use the existing
pt_adjust_global_vcpu_target helper.

Signed-off-by: Roger Pau Monné <roger.pau@citrix.com>
---
Changes since v1:
 - New in this version.