x86/hvm: remove fixed i8259 target
The target vCPU of the i8259 generated interrupts will change
depending on the IO-APIC pin 0, or the configuration of the local APIC
on each vCPU. Also i8259 generated interrupts when delivered using the
virtual wire mode could target multiple vCPUs, as the ExtINT mode
allows for a list of destinations to be configured, like the Fixed
delivery mode.

This at least fixes a bug where the i8259 target vCPU going offline
would prevent further delivery of i8259 generated interrupts, as the
target won't be updated in that case. Also the code in
vlapic_adjust_i8259_target was flawed as it won't take into account
whether vCPUs where offline.

This brings the i8259 emulation closer to how real hardware works, as
allows multiple ExtINT targets, and doesn't latch an i8259 target
which is likely to become stale if not properly updated.

Reported-by: Igor Druzhinin <igor.druzhinin@citrix.com>
Signed-off-by: Roger Pau Monné <roger.pau@citrix.com>