x86/pv: allow reading FEATURE_CONTROL MSR
Linux PV guests will attempt to read the FEATURE_CONTROL MSR, so move
the handling done in VMX code into guest_rdmsr as it can be shared
between PV and HVM guests that way.

Note that there's a slight behavior change and attempting to read the
MSR when no features are available will result in a fault.

Signed-off-by: Roger Pau Monné <roger.pau@citrix.com>
---
Changes since v2:
 - Only allow reading the MSR when there are bits available (different
   than bit 0).

Changes from v1:
 - Move the VMX implementation into guest_rdmsr.