x86/svm: drop writes to BU_CFG on revF chips
We already have special casing to handle reads of this MSR for revF
chips, so do as the comment in svm_msr_read_intercept says and drops
writes. This is in preparation for changing the default MSR write
behavior, which will instead return #GP on not explicitly handled
writes.

Signed-off-by: Roger Pau Monné <roger.pau@citrix.com>
---
Changes since v1:
 - New in this version.