x86/efi: enable MS ABI attribute on clang
Or else the EFI service calls will use the wrong calling convention.

The __ms_abi__ attribute is available on all supported versions of
clang.

Signed-off-by: Roger Pau Monné <roger.pau@citrix.com>
---
Without this a Xen built with clang won't be able to correctly use the
EFI services, leading to weird messages from the firmware. The impact
of this for GCC users is exactly 0, and will fix the build on clang.