autoconf: check endian.h include path
Introduce an autoconf macro to check for the include path of certain
headers that can be different between OSes.

Use such macro to find the correct path for the endian.h header, and
modify the users of endian.h to use the output of such check.

Suggested-by: Ian Jackson <iwj@xenproject.org>
Signed-off-by: Roger Pau Monné <roger.pau@citrix.com>
---
Please re-run autogen after applying.

The biggest risk for this would be some kind of configure or build
failure, and we should be able to catch it in either osstest or the
gitlab build tests.