cirrus-ci: introduce some basic FreeBSD testing
Cirrus-CI supports FreeBSD natively, so introduce some build tests on
FreeBSD.

The Cirrus-CI requires a Github repository in order to trigger the
tests.

A sample run output can be seen at:

https://github.com/royger/xen/runs/1962451343

Note the FreeBSD 11 task fails to build QEMU and is not part of this
patch.

Signed-off-by: Roger Pau Monné <roger.pau@citrix.com>
---
I think this is fine to go in right now, as it's just a meta file for
a test framework, and has 0 chance of breaking anything existing in
Xen.