hvmloadder/rombios: build in freestanding environment
hvmloader and rombios should only use the headers available in the
freestanding environment provided by the compiler. Relying on host
libc headers can cause issues as the bitness of the host is likely
different from the bitness of the binaries (64bit vs 32bit).

Note that on hvmloader the include of inttypes.h is replaced with
stdint.h, since that contains the types required for elfstructs.h.

Remove the requisite to have multilib support and a i386 libc
installed when building on an amd64 host.

Signed-off-by: Roger Pau Monné <roger.pau@citrix.com>